requirements.txt
langchain
langchain-community
openai
python-dotenv
playwright
agentql